# backendCargoFlow/src/main/resources/openapi/trip.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Trip management
  description: API related to management of Trip
paths:
  /trip/create:
    post:
      tags:
        - Trips
      summary: Add a new trip by dispatcher or higher rank
      operationId: addNewTrip
      requestBody:
        description: Trip data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddNewTripRequest'
        required: true
      responses:
        '201':
          description: Trip was created successfully
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/search:
    post:
      tags:
        - Trips
      summary: Search trips with filters and pagination
      operationId: searchTrips
      requestBody:
        description: Trip filters and pagination data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripSearchRequest"
        required: true
      responses:
        '200':
          description: Trips were retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripPageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/{id}:
    get:
      tags:
        - Trips
      summary: Get trip by ID
      operationId: getTrip
      parameters:
        - name: id
          in: path
          description: Trip's ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trip was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '404':
          description: Trip was not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/currentTrip:
    get:
      tags:
        - Trips
      summary: Get current trip for authenticated driver
      operationId: getCurrentTrip
      responses:
        '200':
          description: Current trip was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentTrip'
        '404':
          description: Current trip was not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/completedTrips:
    get:
      tags:
        - Trips
      summary: Get completed trips for authenticated driver
      operationId: getCompletedTrips
      parameters:
        - name: days
          in: query
          description: Number of days in the past. Allowed values are 30, 60 or 90
          required: true
          schema:
            type: integer
            enum:
              - 30
              - 60
              - 90
            example: 30
      responses:
        '200':
          description: Completed driver trips were retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletedTripsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/{tripId}/cancel:
    patch:
      tags:
        - Trips
      summary: Cancel trip
      operationId: cancelTrip
      parameters:
        - name: tripId
          in: path
          description: Trip's ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trip was canceled successfully
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '400':
          description: Trip cannot be canceled
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '404':
          description: Trip was not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'

  /trip/{tripId}/advance-status:
    patch:
      tags:
        - Trips
      summary: Advance driver trip status
      operationId: advanceTripStatus
      parameters:
        - name: tripId
          in: path
          description: Trip's ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trip status was advanced successfully
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '400':
          description: Trip status cannot be advanced
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '404':
          description: Trip was not found for authenticated driver
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
  /trip/dashboard/summary:
    get:
      tags:
        - Trips
      summary: Get trip dashboard summary
      operationId: getTripDashboardSummary
      responses:
        '200':
          description: Trip dashboard summary was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDashboardSummaryResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/GenericApplicationResponse'
components:
  schemas:
    TripStatus:
      type: string
      enum:
        - PLANNED
        - ASSIGNED
        - IN_PROGRESS
        - COMPLETED
        - CANCELED
      example: PLANNED
    CargoType:
      type: string
      enum:
        - GENERAL
        - REFRIGERATED
        - DANGEROUS
        - FRAGILE
    Currency:
      type: string
      enum:
        - EUR
        - RON
        - USD
        - GBP
        - HUF
        - PLN
        - BGN
        - CZK
        - CHF
        - TRY
      example: EUR
    Address:
      type: object
      required:
        - country
        - administrativeArea
        - city
        - streetName
        - streetNumber
        - postalCode
      properties:
        country:
          type: string
          minLength: 2
          maxLength: 50
          example: Romania
        administrativeArea:
          type: string
          minLength: 2
          maxLength: 50
          example: Cluj
        city:
          type: string
          minLength: 2
          maxLength: 50
          example: Cluj-Napoca
        streetName:
          type: string
          minLength: 2
          maxLength: 100
          example: Stefan cel Mare
        streetNumber:
          type: string
          minLength: 1
          maxLength: 15
          example: 52A
        postalCode:
          type: string
          minLength: 2
          maxLength: 15
          example: 123456
        additionalDetails:
          type: string
          maxLength: 250
          nullable: true
          example: Delivery street is one-way. Access only from the main road.
    AddNewTripRequest:
      type: object
      required:
        - pickupAddress
        - deliveryAddress
        - pickupDateTime
        - pickupTimeZone
        - deliveryDateTime
        - deliveryTimeZone
        - cargoType
        - price
        - currency
      properties:
        pickupAddress:
          $ref: '#/components/schemas/Address'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        pickupDateTime:
          type: string
          format: local-date-time
          example: 2026-05-15T12:00:00
        pickupTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Bucharest
        deliveryDateTime:
          type: string
          format: local-date-time
          example: 2026-05-16T19:00:00
        deliveryTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Berlin
        cargoDescription:
          type: string
          minLength: 3
          maxLength: 250
          nullable: true
          example: 10 pallets of electronics
        cargoWeight:
          type: integer
          minimum: 1
          maximum: 24000
          nullable: true
          example: 12345
        cargoVolume:
          type: integer
          minimum: 1
          maximum: 90
          nullable: true
          example: 65
        cargoType:
          $ref: '#/components/schemas/CargoType'
        price:
          type: number
          format: double
          minimum: 0
          example: 1200.5
        currency:
          $ref: '#/components/schemas/Currency'
        additionalInfo:
          type: string
          maxLength: 250
          nullable: true
          example: Driver must call 30 minutes before arrival
    TripSearchRequest:
      type: object
      required:
        - pageNumber
        - pageSize
      properties:
        tripStatusList:
          type: array
          description: If null or empty, trips are not filtered by status
          items:
            $ref: '#/components/schemas/TripStatus'
          example:
            - PLANNED
            - ASSIGNED
        pickupCountries:
          type: array
          description: If null or empty, trips are not filtered by pickup country
          items:
            type: string
            minLength: 2
            maxLength: 50
          example:
            - Romania
            - Germany
        pickupCities:
          type: array
          description: If null or empty, trips are not filtered by pickup city
          items:
            type: string
            minLength: 2
            maxLength: 50
          example:
            - Cluj-Napoca
            - Bucharest
        deliveryCountries:
          type: array
          description: If null or empty, trips are not filtered by delivery country
          items:
            type: string
            minLength: 2
            maxLength: 50
          example:
            - Germany
            - Austria
        deliveryCities:
          type: array
          description: If null or empty, trips are not filtered by delivery city
          items:
            type: string
            minLength: 2
            maxLength: 50
          example:
            - Berlin
            - Vienna
        pickupDateTimeFrom:
          type: string
          format: local-date-time
          example: 2026-05-15T00:00:00
        pickupDateTimeTo:
          type: string
          format: local-date-time
          example: 2026-05-20T23:59:59
        deliveryDateTimeFrom:
          type: string
          format: local-date-time
          example: 2026-05-16T00:00:00
        deliveryDateTimeTo:
          type: string
          format: local-date-time
          example: 2026-05-25T23:59:59
        pageNumber:
          type: integer
          minimum: 0
          example: 0
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
          example: 20
    TripSummary:
      type: object
      required:
        - id
        - tripStatus
        - pickupCountry
        - pickupCity
        - deliveryCountry
        - deliveryCity
        - pickupDateTime
        - pickupTimeZone
        - deliveryDateTime
        - deliveryTimeZone
      properties:
        id:
          type: integer
          format: int64
          example: 15
        tripStatus:
          $ref: '#/components/schemas/TripStatus'
        pickupCountry:
          type: string
          example: Romania
        pickupCity:
          type: string
          example: Cluj-Napoca
        deliveryCountry:
          type: string
          example: Germany
        deliveryCity:
          type: string
          example: Berlin
        pickupDateTime:
          type: string
          format: local-date-time
          example: 2026-05-15T12:00:00
        pickupTimeZone:
          type: string
          example: Europe/Bucharest
        deliveryDateTime:
          type: string
          format: local-date-time
          example: 2026-05-16T19:00:00
        deliveryTimeZone:
          type: string
          example: Europe/Berlin
    TripPageResponse:
      type: object
      required:
        - trips
        - pageNumber
        - pageSize
        - lastPage
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/TripSummary'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        lastPage:
          type: boolean
          example: false
    Trip:
      type: object
      required:
        - id
        - tripStatus
        - pickupAddress
        - deliveryAddress
        - pickupDateTime
        - pickupTimeZone
        - deliveryDateTime
        - deliveryTimeZone
        - cargoType
        - price
        - currency
        - createdBy
      properties:
        id:
          type: integer
          format: int64
          example: 123
        tripStatus:
          $ref: '#/components/schemas/TripStatus'
        pickupAddress:
          $ref: '#/components/schemas/Address'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        pickupDateTime:
          type: string
          format: local-date-time
          example: 2026-05-15T12:00:00
        pickupTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Bucharest
        deliveryDateTime:
          type: string
          format: local-date-time
          example: 2026-05-16T19:00:00
        deliveryTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Berlin
        cargoDescription:
          type: string
          minLength: 3
          maxLength: 250
          nullable: true
          example: 10 pallets of electronics
        cargoWeight:
          type: integer
          minimum: 1
          maximum: 24000
          nullable: true
          example: 12345
        cargoVolume:
          type: integer
          minimum: 1
          maximum: 90
          nullable: true
          example: 65
        cargoType:
          $ref: '#/components/schemas/CargoType'
        price:
          type: number
          format: double
          minimum: 0
          example: 1200.5
        currency:
          $ref: '#/components/schemas/Currency'
        additionalInfo:
          type: string
          maxLength: 250
          nullable: true
          example: Driver must call 30 minutes before arrival
        createdBy:
          type: string
          example: Isac Alexandru
    CurrentTrip:
      type: object
      required:
        - tripId
        - tripStatus
        - pickupAddress
        - deliveryAddress
        - pickupDateTime
        - pickupTimeZone
        - deliveryDateTime
        - deliveryTimeZone
        - cargoType
      properties:
        tripId:
          type: integer
          format: int64
          example: 123
        tripStatus:
          $ref: '#/components/schemas/TripStatus'
        pickupAddress:
          $ref: '#/components/schemas/Address'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        pickupDateTime:
          type: string
          format: local-date-time
          example: 2026-05-15T12:00:00
        pickupTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Bucharest
        deliveryDateTime:
          type: string
          format: local-date-time
          example: 2026-05-16T19:00:00
        deliveryTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Berlin
        cargoDescription:
          type: string
          minLength: 3
          maxLength: 250
          nullable: true
          example: 10 pallets of electronics
        cargoWeight:
          type: integer
          minimum: 1
          maximum: 24000
          nullable: true
          example: 12345
        cargoVolume:
          type: integer
          minimum: 1
          maximum: 90
          nullable: true
          example: 65
        cargoType:
          $ref: '#/components/schemas/CargoType'
        additionalInfo:
          type: string
          maxLength: 250
          nullable: true
          example: Driver must call 30 minutes before arrival
    CompletedTrip:
      type: object
      required:
        - tripId
        - pickupAddress
        - deliveryAddress
        - pickupDateTime
        - pickupTimeZone
        - deliveryDateTime
        - deliveryTimeZone
        - cargoType
      properties:
        tripId:
          type: integer
          format: int64
          example: 123
        pickupAddress:
          $ref: '#/components/schemas/Address'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        pickupDateTime:
          type: string
          format: local-date-time
          example: 2026-05-15T12:00:00
        pickupTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Bucharest
        deliveryDateTime:
          type: string
          format: local-date-time
          example: 2026-05-16T19:00:00
        deliveryTimeZone:
          type: string
          minLength: 3
          maxLength: 50
          example: Europe/Berlin
        cargoDescription:
          type: string
          minLength: 3
          maxLength: 250
          nullable: true
          example: 10 pallets of electronics
        cargoWeight:
          type: integer
          minimum: 1
          maximum: 24000
          nullable: true
          example: 12345
        cargoVolume:
          type: integer
          minimum: 1
          maximum: 90
          nullable: true
          example: 65
        cargoType:
          $ref: '#/components/schemas/CargoType'
        additionalInfo:
          type: string
          maxLength: 250
          nullable: true
          example: Driver must call 30 minutes before arrival
    CompletedTripsResponse:
      type: object
      required:
        - trips
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/CompletedTrip'
    TripDashboardSummaryItem:
      type: object
      required:
        - tripStatus
        - value
      properties:
        tripStatus:
          $ref: '#/components/schemas/TripStatus'
        value:
          type: integer
          format: int64
          example: 12
    TripDashboardSummaryResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TripDashboardSummaryItem'